+++ lib/rules/no-top-level-side-effects.ts
@@ -5,7 +5,8 @@ import type {
   CallExpression,
   NewExpression,
   AssignmentExpression,
-  VariableDeclarator
+  VariableDeclarator,
+  Identifier
 } from 'estree';
 
 import {IsCommonJs, isTopLevel} from '../helpers';
@@ -50,9 +51,36 @@ const disallowedSideEffect = {
 };
 
 function isCallTo(expression: CallExpression, name: string): boolean {
-  return (
-    expression.callee.type === 'Identifier' && expression.callee.name === name
-  );
+  if (expression.callee.type === 'Identifier') {
+    return expression.callee.name === name;
+  }
+
+  if (expression.callee.type === 'MemberExpression') {
+    let expr = expression.callee;
+    const id: string[] = [];
+    while (true) {
+      if (expr.computed) {
+        return false;
+      }
+
+      const property = expr.property as Identifier; // type-coverage:ignore-line
+      id.push(property.name);
+
+      if (expr.object.type === 'MemberExpression') {
+        expr = expr.object;
+      } else if (expr.object.type === 'Identifier') {
+        id.push(expr.object.name);
+        break;
+      } else {
+        return false;
+      }
+    }
+
+    id.reverse();
+    return name === id.join('.');
+  }
+
+  return false;
 }
 
 function isDestructuring(declaration: VariableDeclarator): boolean {
@@ -164,7 +192,7 @@ export const noTopLevelSideEffects: Rule.RuleModule = {
     }
   },
   create: (context) => {
-    const provided: Partial<Options> = context.options[0]; // type-coverage:ignore-line
+    const [provided] = context.options as Partial<Options>[]; // type-coverage:ignore-line
 
     const options: Options = {
       allowDerived: provided?.allowDerived || allowDerivedOption.default,