+++ CHANGELOG.md
@@ -10,7 +10,13 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- ([#1506]) Add support for `using` and `await using`.
+
+## [3.5.2] - 2025-05-11
+
+- ([#1474]) Add support for Node.js v24.
+- ([#1475]) Add support for member expression in the `allowedCalls` option of
+  the `no-top-level-side-effects` rule.
 
 ## [3.5.1] - 2025-03-21
 
@@ -213,3 +219,6 @@ Versioning].
 
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
+[#1474]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1474
+[#1475]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1475
+[#1506]: https://github.com/ericcornelissen/eslint-plugin-top/pull/1506